Moving websockets_base.py into websockets_common folder
